requirements.txt
pip==22.2.2
setuptools==65.4.0
wheel==0.37.1
selenium==4.4.3
pytest==7.1.3
allure-pytest==2.9.45
